storyline4_pipeline/requirements.txt
pandas>=1.5.0
openpyxl>=3.0.10
numpy>=1.23.0
matplotlib>=3.6.0
PyYAML>=6.0
